+++ src/shared/auth0/client.ts
